% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_get_regex.R
\name{ojo_get_flag_regex}
\alias{ojo_get_flag_regex}
\title{Return OJO Regex for a given flag}
\usage{
ojo_get_flag_regex(flag = NA)
}
\arguments{
\item{flag}{The flag you want to get the regex pattern for}
}
\value{
A string of regex
}
\description{
This function returns the regex string for a given flag
}
