% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_get_regex.R
\name{ojo_get_statute_regex}
\alias{ojo_get_statute_regex}
\title{Return OJO Regex for a given statute}
\usage{
ojo_get_statute_regex(statute = NA)
}
\arguments{
\item{statute}{The statute you want to get the regex pattern for}
}
\value{
A string of regex
}
\description{
This function returns the regex string for a given statute
}
