% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_regex_cats.R
\docType{data}
\name{ojo_regex_cats}
\alias{ojo_regex_cats}
\title{OJO Regex Categories dataset}
\format{
A data frame with X rows and 13 columns:
\describe{
\item{in_ojoregex}{Description of in_ojoregex}
\item{clean_charge_description}{Description of clean_charge_description}
\item{category}{Description of category}
\item{subcategory}{Description of subcategory}
\item{title}{Description of title}
\item{statutes}{Description of statutes}
\item{chapter}{Description of chapter}
\item{gist}{Description of gist}
\item{description}{Description of description}
\item{statute_link}{Description of statute_link}
\item{cf_cm}{Description of cf_cm}
\item{cf_cm_notes}{Description of cf_cm_notes}
\item{outdated}{Description of outdated}
\item{notes}{Description of notes}
}
}
\usage{
ojo_regex_cats
}
\description{
OJO Regex Categories dataset
}
\keyword{datasets}
