% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_regex_flags.R
\docType{data}
\name{ojo_regex_flags}
\alias{ojo_regex_flags}
\title{OJO Regex Flags Dataset}
\format{
A data frame with 139 rows and 8 columns:
\describe{
\item{flag}{Description of column1}
\item{regex}{Description of column2}
\item{group}{Description of column3}
\item{criminal_or_civil}{Description of column4}
\item{genre}{Description of column5}
\item{word_boundary}{Description of column6}
\item{examples}{Description of column7}
\item{notes}{Description of column8}
...
}
}
\usage{
ojo_regex_flags
}
\description{
OJO Regex Flags Dataset
}
\keyword{datasets}
