% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_regex.R
\docType{data}
\name{ojo_regex_unhoused}
\alias{ojo_regex_unhoused}
\title{Unhoused / Homeless address regex}
\format{
An object of class \code{character} of length 1.
}
\usage{
ojo_regex_unhoused
}
\description{
Detects whether an address is likely to indicate "homeless" / "unhoused" / "NA", etc.
}
\keyword{datasets}
