% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clean.R
\name{regex_pre_clean}
\alias{regex_pre_clean}
\title{Pre clean charge descriptions to be matched}
\usage{
regex_pre_clean(count_as_filed)
}
\arguments{
\item{text}{}
}
\description{
Pre clean charge descriptions to be matched
}
