% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_datasets.R
\name{update_datasets}
\alias{update_datasets}
\title{Update ojoRegex datasets to match the Big Ol Spreadsheet}
\usage{
update_datasets(email = NULL)
}
\arguments{
\item{email}{The email address to use for Google authentication.}
}
\value{
A cleaned and categorized dataset with charge descriptions in the specified column, along with any additional columns present in the original dataset.
}
\description{
This function reads the Big Ol Spreadsheet and saves the data to the package data folder. It's not exported / available to package users becuase it's intended to be used during development only, so don't run this unless you're developing the package!!
}
\examples{
\dontrun{
update_datasets("abell@okpolicy.org")
}
}
