% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clean.R
\name{regex_pre_clean}
\alias{regex_pre_clean}
\title{Pre clean charge descriptions to be matched}
\usage{
regex_pre_clean(count_as_filed)
}
\arguments{
\item{text}{A character vector containing the charge descriptions to be pre-cleaned.}
}
\value{
A character vector with pre-cleaned charge descriptions.
}
\description{
This function pre-cleans charge descriptions to be matched by removing specific patterns that are not relevant for matching. It removes phrases like "in concert with" from the end of the charge descriptions.
}
\examples{
\dontrun{
# Example usage
clean_text <- regex_pre_clean("TAXS, FAIL TO DISPLAY TAX STAMP ON CDS IN CONCERT W/J POOLE")
clean_text
}
}
